refactor: 브랜드,상품 상위에는 카탈로그가 존재
