--- apps/commerce-api/src/main/java/com/loopers/domain/brand/BrandJpaRepository.java
@@ -1,4 +1,0 @@
-package com.loopers.domain.brand;
-
-public interface BrandJpaRepository {
-}
