+++ apps/commerce-api/src/main/java/com/loopers/domain/catalog/brand/BrandJpaRepository.java
@@ -1,0 +1,4 @@
+package com.loopers.domain.catalog.brand;
+
+public interface BrandJpaRepository {
+}
