+++ apps/commerce-api/src/main/java/com/loopers/domain/catalog/product/embeded/ProductName.java
@@ -1,4 +1,4 @@
-package com.loopers.domain.product.embeded;
+package com.loopers.domain.catalog.product.embeded;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
